Refactor customers RESTAPI handling into a handler (removes routing logic from controller)
The handler is responsible of converting a HTTP request into a Response
object. How this is done is left to the handler. This implenents a
routing handler for the customer handler which is able to resolve its
actions via a routing definition (using the symfony routing component
for sub-path, method and parameter matching).